scripts/article_embedding/requirements.txt
boto3
openai
psycopg2
python-dotenv   # If using .env files for credentials
tiktoken
nltk